move vec_merge to utils
